etl/requirements.txt
psycopg2-binary==2.9
pydantic==1.10.2
backoff==2.2.1
elasticsearch==8.5.3